Seperate rich to its own module
